[ALPS06729315] mtk_vcodec: add CID for encode feature[1/2]
1 add CID for multiple reference and TSVC
2 add CID for mlvec featruelist:
   prepend hdr, inputqueue mode, slice header spacing, extend_SAR,
  dynamic tsvc and LTR

MTK-Commit-Id: d84a906537d9a660e087e7b75f47b3fb008d09c8

CR-Id: ALPS06729315
Feature: Video Driver

+++ drivers/media/platform/mtk-vcodec/venc_vcu_if.h
@@ -32,7 +32,7 @@ struct venc_vcu_inst {
 	int failure;
 	int bs_size;
 	int is_key_frm;
-	unsigned int inst_addr;
+	__u64 inst_addr;
 	void *vsi;
 	enum ipi_id id;
 	struct mtk_vcodec_ctx *ctx;
